Resolve "Canoscan LiDE 220 aborts before end of 4800dpi colour scan"

Closes #580

See merge request sane-project/backends!697

+++ backend/genesys/gl841.cpp
@@ -978,7 +978,7 @@ dummy \ scanned lines
     dev->session = session;
 
     dev->total_bytes_read = 0;
-    dev->total_bytes_to_read = session.output_line_bytes_requested * session.params.lines;
+    dev->total_bytes_to_read = (size_t)session.output_line_bytes_requested * (size_t)session.params.lines;
     if (session.use_host_side_gray) {
         dev->total_bytes_to_read /= 3;
     }
